execution-simulator/requirements.txt

fastapi==0.108.0
uvicorn==0.25.0
nats-py==2.1.4
python-dotenv==1.0.0